Enhance JIT and Transform Functionality
- Updated `remap_dict_indices_by_chunk` to use `cuda::stream_ref` for improved stream handling.
- Modified `get_kernel` and related functions to accept a `kernel_entry` parameter for better kernel management.
- Introduced `row_accessor` for efficient row-wise access in JIT transformations, supporting both null-aware and non-null-aware operations.
- Added `list_row` and `struct_row` types to handle nested structures in JIT transformations.
- Implemented lazy row access in transform kernels to optimize memory usage and performance.
- Enhanced hashing tests to include JIT variants and added tests for mixed nullable columns and nested structures.
- Updated transform tests to validate new row accessor functionality and ensure compatibility with existing UDFs.

+++ cpp/include/cudf/hashing/detail/hashing.hpp
@@ -52,7 +52,7 @@ std::unique_ptr<column> sha512(table_view const& input,
                                rmm::device_async_resource_ref mr);
 
 std::unique_ptr<column> xxhash_32(table_view const& input,
-                                  uint64_t seed,
+                                  uint32_t seed,
                                   cuda::stream_ref,
                                   rmm::device_async_resource_ref mr);
